Providing a work-around-ish solution to #1565.
- Avoids using a default implementation when the defaultImpl anntotation
  should not apply.
