+++ src/main/java/com/fasterxml/jackson/databind/jsontype/impl/StdTypeResolverBuilder.java
@@ -1,12 +1,15 @@
 package com.fasterxml.jackson.databind.jsontype.impl;
 
+import java.util.ArrayList;
 import java.util.Collection;
+import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonTypeInfo;
 
 import com.fasterxml.jackson.databind.*;
 import com.fasterxml.jackson.databind.annotation.NoClass;
 import com.fasterxml.jackson.databind.cfg.MapperConfig;
+import com.fasterxml.jackson.databind.introspect.AnnotatedClass;
 import com.fasterxml.jackson.databind.jsontype.*;
 
 /**
@@ -125,8 +128,13 @@ public class StdTypeResolverBuilder
                      || (_defaultImpl == NoClass.class)) {
                 defaultImpl = config.getTypeFactory().constructType(_defaultImpl);
             } else {
-                defaultImpl = config.getTypeFactory()
-                    .constructSpecializedType(baseType, _defaultImpl);
+                if (!baseType.getRawClass().isAssignableFrom(_defaultImpl) && isBaseActuallySpecializedType(config, baseType)) {
+                    defaultImpl = null;
+                }
+                else {
+                    defaultImpl = config.getTypeFactory()
+                        .constructSpecializedType(baseType, _defaultImpl);
+                }
             }
         }
 
@@ -147,6 +155,36 @@ public class StdTypeResolverBuilder
                     _typeProperty, _typeIdVisible, defaultImpl);
         }
         throw new IllegalStateException("Do not know how to construct standard type serializer for inclusion type: "+_includeAs);
+    }
+
+    private boolean isBaseActuallySpecializedType(DeserializationConfig config, JavaType baseType) {
+        // 12-Jun-2017, slobo: There is a chance that we're deserializing
+        // a specific class part of a polymorphic hierarchy. In this case
+        // baseType is actually a specific type, but the _defaultImpl comes
+        // from the real base type. For example:
+        //
+        //     B @JsonTypeInfo(defaultImpl=S2)
+        //  /--+--\
+        //  S1    S2
+        //  baseType = S1
+        //  _defaultImpl = S2
+        //
+        // To detect this scenario we'll check whether _defaultImpl and
+        // baseType share a common superclass or super superclass, etc.
+        JavaType defaultType = config.getTypeFactory().constructType(_defaultImpl);
+        for (JavaType current = defaultType; current != null; current = current.getSuperClass())
+        {
+            AnnotatedClass annotatedClass = AnnotatedClass.construct(current, config);
+            if (annotatedClass.hasAnnotation(JsonTypeInfo.class)) {
+                if (annotatedClass.getRawType().isAssignableFrom(baseType.getRawClass())) {
+                    return true;
+                }
+            }
+            else {
+                break; // ignore any classes above the JsonTypeInfo annotation
+            }
+        };
+        return false;
     }
 
     /*
